VYZN: Matrix Messaging Integration (#455)
* install matrix-widget-api

* bldrs to vyzn integration wip

* test out routing

* refactor code, use store instead of router

* changes for codeReview

* move WidgetApi into react context, use navigation() instead of the detour in the store, add event emitter

* seperate events for selection an unselection of elements

* select multiple elements

* remove code comments

* changes from codereview

* VYZN-806: Tests for Bldrs integration

* fix issue after merge

* fix new linting issues after merge

* trigger preview deployment

* stabilize cypress integration tests

* cadview test fix

* fix yarn lock conflict

* parameters bug fix & adding test case for it

* added matrix-widget-api to the lock file

+++ src/Containers/CadView.jsx
@@ -28,7 +28,7 @@ import {handleBeforeUnload} from '../utils/event'
  * Experimenting with a global. Just calling #indexElement and #clear
  * when new models load.
  */
-const searchIndex = new SearchIndex()
+export const searchIndex = new SearchIndex()
 let count = 0
 
 
@@ -74,6 +74,7 @@ export default function CadView({
   const setModelStore = useStore((state) => state.setModelStore)
   const setSelectedElement = useStore((state) => state.setSelectedElement)
   const setSelectedElements = useStore((state) => state.setSelectedElements)
+  const selectedElements = useStore((state) => state.selectedElements)
   const setViewerStore = useStore((state) => state.setViewerStore)
   const snackMessage = useStore((state) => state.snackMessage)
   const [modelReady, setModelReady] = useState(false)
@@ -103,6 +104,17 @@ export default function CadView({
   useEffect(() => {
     onSearchParams()
   }, [searchParams])
+
+
+  useEffect(() => {
+    if (Array.isArray(selectedElements)) {
+      if (selectedElements.length > 0) {
+        selectItemsInScene(selectedElements.map((id) => parseInt(id)))
+      } else {
+        viewer?.IFC.unpickIfcItems()
+      }
+    }
+  }, [selectedElements])
 
 
   // Watch for path changes within the model.
@@ -339,7 +351,7 @@ export default function CadView({
         throw new Error('IllegalState: empty search query')
       }
       const resultIDs = searchIndex.search(query)
-      selectItemsInScene(resultIDs)
+      setSelectedElements(resultIDs)
       setDefaultExpandedElements(resultIDs.map((id) => `${id}`))
       Privacy.recordEvent('search', {
         search_term: query,
@@ -369,8 +381,10 @@ export default function CadView({
 
   /** Unpick active scene elts and remove clip planes. */
   function unSelectItems() {
-    viewer.IFC.unpickIfcItems()
-    viewer.clipper.deleteAllPlanes()
+    if (viewer) {
+      viewer.IFC.unpickIfcItems()
+      viewer.clipper.deleteAllPlanes()
+    }
     resetState()
     const repoFilePath = modelPath.gitpath ? modelPath.getRepoPath() : modelPath.filepath
     window.removeEventListener('beforeunload', handleBeforeUnload)
@@ -384,7 +398,6 @@ export default function CadView({
    * @param {Array} resultIDs Array of expressIDs
    */
   async function selectItemsInScene(resultIDs) {
-    setSelectedElements(resultIDs.map((id) => `${id}`))
     try {
       await viewer.pickIfcItemsByID(0, resultIDs, true)
     } catch (e) {
@@ -410,10 +423,9 @@ export default function CadView({
       debug().error(`CadView#onElementSelect(${expressId}) missing in table:`, elementsById)
       return
     }
-    await selectItemsInScene([expressId])
     const pathIds = computeElementPathIds(lookupElt, (elt) => elt.expressID)
     setExpandedElements(pathIds.map((n) => `${n}`))
-    setSelectedElements(`${expressId}`)
+    setSelectedElements([`${expressId}`])
     const props = await viewer.getProperties(0, expressId)
     setSelectedElement(props)
     return pathIds
